+++ src/graphics/tilemap.h
@@ -19,11 +19,11 @@ typedef struct
 } TilemapLayer;
 
 void tilemap_init(Tilemap *tilemap, Graphics *graphics, TextureEntry *tileset,
-                 TransformEntry transform, int map_w, int map_h, int layers,
-                 u32 *map_data);
+                  TransformEntry transform, int map_w, int map_h, int layers,
+                  i32 *map_data);
 void tilemap_free(Tilemap *tilemap, Graphics *graphics);
 
 void tilemap_render(Tilemap *tilemap, mat4s camera, int layer,
                     WGPURenderPassEncoder pass);
 void tilemap_set_tile(Tilemap *tilemap, Graphics *graphics, int x, int y,
-                      int layer, u16 tile);
+                      int layer, i32 tile);
